+++ bundles/org.palladiosimulator.analyzer.slingshot.singlestatesimulation.application/launchconfig/headless-SSS-wtf.launch
@@ -86,6 +86,8 @@
         <setEntry value="lpg.runtime.java@default:default"/>
         <setEntry value="net.bytebuddy.byte-buddy-agent@default:default"/>
         <setEntry value="net.bytebuddy.byte-buddy@default:default"/>
+        <setEntry value="org.antlr.runtime*3.2.0.v201101311130@default:default"/>
+        <setEntry value="org.antlr.runtime*4.7.2.v20200218-0804@default:default"/>
         <setEntry value="org.aopalliance@default:default"/>
         <setEntry value="org.apache.ant@default:default"/>
         <setEntry value="org.apache.batik.bridge@default:default"/>
@@ -97,20 +99,25 @@
         <setEntry value="org.apache.batik.dom*1.6.1.v201505192100@default:default"/>
         <setEntry value="org.apache.batik.dom*1.9.1.v20190730-1743@default:default"/>
         <setEntry value="org.apache.batik.dom.svg@default:default"/>
+        <setEntry value="org.apache.batik.ext.awt*1.6.0.v201011041432@default:default"/>
+        <setEntry value="org.apache.batik.ext.awt*1.9.1.v20190730-1743@default:default"/>
         <setEntry value="org.apache.batik.i18n*1.14.0.v20210324-0332@default:default"/>
         <setEntry value="org.apache.batik.i18n*1.9.1.v20180227-1645@default:default"/>
         <setEntry value="org.apache.batik.parser@default:default"/>
+        <setEntry value="org.apache.batik.pdf@default:default"/>
         <setEntry value="org.apache.batik.svggen*1.6.0.v201011041432@default:default"/>
         <setEntry value="org.apache.batik.svggen*1.9.1.v20190730-1743@default:default"/>
         <setEntry value="org.apache.batik.transcoder@default:default"/>
         <setEntry value="org.apache.batik.util*1.14.0.v20210324-0332@default:default"/>
         <setEntry value="org.apache.batik.util*1.6.0.v201011041432@default:default"/>
         <setEntry value="org.apache.batik.util*1.9.1.v20180703-1553@default:default"/>
+        <setEntry value="org.apache.batik.util.gui@default:default"/>
         <setEntry value="org.apache.batik.xml*1.6.0.v201011041432@default:default"/>
         <setEntry value="org.apache.batik.xml*1.9.1.v20190730-1743@default:default"/>
         <setEntry value="org.apache.commons.cli@default:default"/>
         <setEntry value="org.apache.commons.codec@default:default"/>
         <setEntry value="org.apache.commons.collections15@default:default"/>
+        <setEntry value="org.apache.commons.io@default:default"/>
         <setEntry value="org.apache.commons.jxpath@default:default"/>
         <setEntry value="org.apache.commons.lang@default:default"/>
         <setEntry value="org.apache.commons.logging.fragment@default:false"/>
@@ -123,6 +130,9 @@
         <setEntry value="org.apache.log4j@default:default"/>
         <setEntry value="org.apache.xalan@default:default"/>
         <setEntry value="org.apache.xmlgraphics@default:default"/>
+        <setEntry value="org.apiguardian@default:default"/>
+        <setEntry value="org.bouncycastle.bcpg@default:default"/>
+        <setEntry value="org.bouncycastle.bcprov@default:default"/>
         <setEntry value="org.eclipse.ant.core@default:default"/>
         <setEntry value="org.eclipse.ant.launching@default:default"/>
         <setEntry value="org.eclipse.ant.ui@default:default"/>
@@ -259,7 +269,10 @@
         <setEntry value="org.eclipse.gmf.runtime.gef.ui@default:default"/>
         <setEntry value="org.eclipse.gmf.runtime.notation@default:default"/>
         <setEntry value="org.eclipse.help@default:default"/>
+        <setEntry value="org.eclipse.jdt.annotation*2.2.600.v20200408-1511@default:default"/>
         <setEntry value="org.eclipse.jdt.apt.core@default:default"/>
+        <setEntry value="org.eclipse.jdt.compiler.apt@default:false"/>
+        <setEntry value="org.eclipse.jdt.compiler.tool@default:false"/>
         <setEntry value="org.eclipse.jdt.core.manipulation@default:default"/>
         <setEntry value="org.eclipse.jdt.core@default:default"/>
         <setEntry value="org.eclipse.jdt.debug.ui@default:default"/>
@@ -344,19 +357,25 @@
         <setEntry value="org.eclipse.xtend.standalone@default:default"/>
         <setEntry value="org.eclipse.xtend.typesystem.emf@default:default"/>
         <setEntry value="org.eclipse.xtend@default:default"/>
+        <setEntry value="org.eclipse.xtext.builder.standalone@default:default"/>
         <setEntry value="org.eclipse.xtext.builder@default:default"/>
         <setEntry value="org.eclipse.xtext.common.types.ui@default:default"/>
         <setEntry value="org.eclipse.xtext.common.types@default:default"/>
         <setEntry value="org.eclipse.xtext.ecore@default:default"/>
+        <setEntry value="org.eclipse.xtext.generator@default:default"/>
         <setEntry value="org.eclipse.xtext.ide@default:default"/>
+        <setEntry value="org.eclipse.xtext.junit4@default:default"/>
         <setEntry value="org.eclipse.xtext.smap@default:default"/>
+        <setEntry value="org.eclipse.xtext.testing@default:default"/>
         <setEntry value="org.eclipse.xtext.ui.codetemplates.ide@default:default"/>
         <setEntry value="org.eclipse.xtext.ui.codetemplates.ui@default:default"/>
         <setEntry value="org.eclipse.xtext.ui.codetemplates@default:default"/>
         <setEntry value="org.eclipse.xtext.ui.shared@default:default"/>
+        <setEntry value="org.eclipse.xtext.ui.testing@default:default"/>
         <setEntry value="org.eclipse.xtext.ui@default:default"/>
         <setEntry value="org.eclipse.xtext.util@default:default"/>
         <setEntry value="org.eclipse.xtext.xbase.ide@default:default"/>
+        <setEntry value="org.eclipse.xtext.xbase.junit@default:default"/>
         <setEntry value="org.eclipse.xtext.xbase.lib@default:default"/>
         <setEntry value="org.eclipse.xtext.xbase@default:default"/>
         <setEntry value="org.eclipse.xtext.xtext.generator@default:default"/>
@@ -465,6 +484,7 @@
         <setEntry value="org.palladiosimulator.analyzer.slingshot.eventdriver.annotationprocessor@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.eventdriver@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.initialisedsimulation.application@default:default"/>
+        <setEntry value="org.palladiosimulator.analyzer.slingshot.initialisedsimulation@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.monitor.data@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.monitor.processingtype.aggregator@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.monitor.processingtype.feedthrough@default:default"/>
@@ -475,7 +495,6 @@
         <setEntry value="org.palladiosimulator.analyzer.slingshot.snapshot@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.stateexploration.converter@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.stateexploration.data@default:default"/>
-        <setEntry value="org.palladiosimulator.analyzer.slingshot.stateexploration@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.ui.events@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.workflow.events@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.workflow@default:default"/>
