+++ bundles/org.palladiosimulator.analyzer.slingshot.singlestatesimulation.application/launchconfig/headless-SSS.launch
@@ -462,6 +462,7 @@
         <setEntry value="org.palladiosimulator.analyzer.slingshot.eventdriver.annotationprocessor@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.eventdriver@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.initialisedsimulation.application@default:default"/>
+        <setEntry value="org.palladiosimulator.analyzer.slingshot.initialisedsimulation@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.monitor.data@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.monitor.processingtype.aggregator@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.monitor.processingtype.feedthrough@default:default"/>
@@ -472,7 +473,6 @@
         <setEntry value="org.palladiosimulator.analyzer.slingshot.snapshot@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.stateexploration.converter@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.stateexploration.data@default:default"/>
-        <setEntry value="org.palladiosimulator.analyzer.slingshot.stateexploration@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.ui.events@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.workflow.events@default:default"/>
         <setEntry value="org.palladiosimulator.analyzer.slingshot.workflow@default:default"/>
