+++ bundles/org.palladiosimulator.analyzer.slingshot.singlestatesimulation.application/src/org/palladiosimulator/analyzer/slingshot/initialisedsimulation/application/InitialisedSimulationApplication.java
@@ -14,9 +14,9 @@ import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.emf.ecore.resource.impl.ResourceSetImpl;
 import org.eclipse.equinox.app.IApplication;
 import org.eclipse.equinox.app.IApplicationContext;
-import org.palladiosimulator.analyzer.slingshot.initialisedsimulation.explorer.configuration.ExplorationWorkflowConfiguration;
-import org.palladiosimulator.analyzer.slingshot.initialisedsimulation.explorer.configuration.InitialiseSimulationRootJob;
-import org.palladiosimulator.analyzer.slingshot.initialisedsimulation.explorer.configuration.InitialiseSimulationWorkflowConfiguration;
+import org.palladiosimulator.analyzer.slingshot.initialisedsimulation.configuration.ExplorationWorkflowConfiguration;
+import org.palladiosimulator.analyzer.slingshot.initialisedsimulation.configuration.InitialiseSimulationWorkflowConfiguration;
+import org.palladiosimulator.analyzer.slingshot.initialisedsimulation.jobs.InitialiseSimulationRootJob;
 import org.palladiosimulator.experimentautomation.application.ExperimentApplication;
 import org.palladiosimulator.experimentautomation.application.tooladapter.abstractsimulation.AbstractSimulationConfigFactory;
 import org.palladiosimulator.experimentautomation.application.tooladapter.slingshot.model.SlingshotConfiguration;
