+++ bundles/org.palladiosimulator.analyzer.slingshot.stateexploration.converter/src/org/palladiosimulator/analyzer/slingshot/converter/data/SLO.java
@@ -1,5 +1,6 @@
 package org.palladiosimulator.analyzer.slingshot.converter.data;
 
+@Deprecated
 public record SLO(String id, String name, String specificationId, Number lowerThreshold, Number upperThreshold) {
 	
 }
