+++ bundles/org.palladiosimulator.analyzer.slingshot.stateexploration/META-INF/MANIFEST.MF
@@ -7,9 +7,9 @@ Automatic-Module-Name: org.palladiosimulator.analyzer.slingshot.stateexploration
 Bundle-ClassPath: lib/json-20240303.jar,
  .
 Bundle-RequiredExecutionEnvironment: JavaSE-17
-Export-Package: org.palladiosimulator.analyzer.slingshot.initialisedsimulation.explorer,
- org.palladiosimulator.analyzer.slingshot.initialisedsimulation.explorer.configuration,
- org.palladiosimulator.analyzer.slingshot.initialisedsimulation.explorer.planning,
+Export-Package: org.palladiosimulator.analyzer.slingshot.initialisedsimulation,
+ org.palladiosimulator.analyzer.slingshot.initialisedsimulation.configuration,
+ org.palladiosimulator.analyzer.slingshot.initialisedsimulation.jobs,
  org.palladiosimulator.analyzer.slingshot.initialisedsimulation.providers
 Require-Bundle: org.palladiosimulator.edp2,
  org.palladiosimulator.analyzer.slingshot.snapshot.data,
