+++ bundles/org.palladiosimulator.analyzer.slingshot.stateexploration/src/org/palladiosimulator/analyzer/slingshot/initialisedsimulation/configuration/ExplorationWorkflowConfiguration.java
@@ -1,4 +1,4 @@
-package org.palladiosimulator.analyzer.slingshot.initialisedsimulation.explorer.configuration;
+package org.palladiosimulator.analyzer.slingshot.initialisedsimulation.configuration;
 
 
 import java.util.Map;
